digital-credentials/digital-credential-client-allows-protocol.https.html

<!DOCTYPE html>
<title>Digital Credential Client Allows Protocol method.</title>
<link rel="help" href="https://w3c-fedid.github.io/digital-credentials/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body></body>
<script type="module">

  const invalidIdentifiers = ["", " ", "a b", "a,b", "AB", "a.B", "a.b", " foo-bar", " foo-bar ", "foo-bar ", "fooBar"];

  promise_test(async (t) => {
    for (const invalidIdentifier of invalidIdentifiers) {
      const result = DigitalCredential.allowsProtocol(invalidIdentifier);
      assert_equals(result, false, `Incorrect result for ${invalidIdentifier}.`);
    }
  }, "Client does not allow invalid protocol identifiers.");

</script>
